+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/k8-automation/compare/0.3.0...HEAD
 
+### Added
+
+-   Set ATOMIST_ENVIRONMENT variable in deployment pod container [#7][7]
+
+[7]: https://github.com/atomist/k8-automation/issues/7
+
 ## [0.3.0][] - 2018-03-01
 
 [0.3.0]: https://github.com/atomist/k8-automation/compare/0.2.6...0.3.0